// src/api/openapi/models/homework-detail-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Zeroone API v1
 * v1
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { EvaluationDetailResponseDto } from './evaluation-detail-response-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { ImageDto } from './image-dto';

export interface HomeworkDetailResponseDto {
    'homeworkId'?: number;
    'homeworkStatus'?: HomeworkDetailResponseDtoHomeworkStatusEnum;
    'submissionTime'?: string;
    'submitterId'?: number;
    /**
     * 그룹스터디 신청자 역할
     */
    'submitterApplyRole'?: HomeworkDetailResponseDtoSubmitterApplyRoleEnum;
    'submitterNickname'?: string;
    'submitterProfileImage'?: ImageDto;
    'homeworkTextContent'?: string;
    'homeworkLink'?: string;
    'evaluation'?: EvaluationDetailResponseDto;
}

export const HomeworkDetailResponseDtoHomeworkStatusEnum = {
    EvaluationCompleted: 'EVALUATION_COMPLETED',
    Submitted: 'SUBMITTED',
    NotSubmitted: 'NOT_SUBMITTED'
} as const;

export type HomeworkDetailResponseDtoHomeworkStatusEnum = typeof HomeworkDetailResponseDtoHomeworkStatusEnum[keyof typeof HomeworkDetailResponseDtoHomeworkStatusEnum];
export const HomeworkDetailResponseDtoSubmitterApplyRoleEnum = {
    Participant: 'PARTICIPANT',
    Leader: 'LEADER'
} as const;

export type HomeworkDetailResponseDtoSubmitterApplyRoleEnum = typeof HomeworkDetailResponseDtoSubmitterApplyRoleEnum[keyof typeof HomeworkDetailResponseDtoSubmitterApplyRoleEnum];